algolegacy/requirements.txt
beaker-pyteal>=0.9.1
algokit-utils>=1.2.0
pyteal>=0.20.1
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
